Use React as peer dependency, update Readme syntax
Using peer dependencies seems to be the convention for utilizing the same
instance of an imported third party library.

Deprecated original API since React no longer needs to be passed as an
argument.  Also passing options as the first arg seems odd.

Updated README to reflect this change along with using more es6-esk
syntax for doing the export (since this is a react module).